requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
#tensorflow-macos==2.10.1; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
#tensorflow==2.10.1; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow==2.17.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips

pytest

fastapi
uvicorn
pydantic
pillow
python-multipart
